Переработать архитектуру. Создайть интерфейсы FilmStorage и UserStorage, в которых определены методы добавления, удаления и модификации объектов. Создать классы InMemoryFilmStorage и InMemoryUserStorage.
