+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,0 +1,7 @@
+package ru.yandex.practicum.filmorate.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class FilmService {
+}
