+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,0 +1,7 @@
+package ru.yandex.practicum.filmorate.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+}
